requirements.txt
selenium==4.10.0
webdriver-manager==3.8.5
playwright==1.34.0
openai==0.27.8
python-dotenv==1.0.0
pandas==2.0.3
gspread==5.10.0
oauth2client==4.1.3
schedule==1.2.0
pytz==2023.3
matplotlib==3.7.2
flask==2.3.3
requests==2.31.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
stripe==6.1.0
python-multipart==0.0.6 